It now also compiles for `unknown` targets. But testing still produces errors.
