+++ src/arch/unknown/intrin/mod.rs
@@ -1,3 +1,5 @@
+mod merge;
+
 pub mod prelude {
-    
+    pub use super::merge::*;
 }
