+++ src/arch/x86/mod.rs
@@ -1,2 +1,3 @@
 pub mod intrin;
 pub mod vecs;
+pub mod vec_patterns;
