+++ src/arch/x86/vec_patterns_gen.py
@@ -27,34 +27,10 @@ print("""// This file is part of faster, the SIMD library for humans.
 // This file is machine-generated. See vec_patterns_gen.py for more info.""")
 
 print("""
-use vecs::*;
-use core_or_std::mem::transmute;
-use stdsimd::vendor::*;
-
-/// Constructors which may be used to instantiate vectors with patterned data.
-pub trait Pattern : Packed {
-    /// Return a vector whose first `Self::WIDTH / 2` elements are `hi`, and
-    /// whose last `Self::WIDTH / 2` elements are `lo`.
-    fn halfs(hi: Self::Scalar, lo: Self::Scalar) -> Self;
-
-    /// Return a vector containing `hi` at every even index, and lo at every odd
-    /// index.
-    fn interleave(hi: Self::Scalar, lo: Self::Scalar) -> Self;
-
-    /// Return a vector whose first `off` elements are `hi`, and whose last
-    /// `Self::WIDTH - off` elements are `lo`.
-    fn partition(hi: Self::Scalar, lo: Self::Scalar, off: usize) -> Self;
-
-    /// Return a vector whose first `off` elements are memset to 0x00, and whose
-    /// last `Self::WIDTH - off` elements are memset to 0xFF.
-    fn partition_mask(off: usize) -> Self;
-
-    /// Return a vector made entirely of ones.
-    fn ones() -> Self;
-
-    /// Return a vector made entirely of zeroes.
-    fn zeroes() -> Self;
-}
+use crate::arch::current::vecs::*;
+use crate::std::mem::transmute;
+use crate::vecs::*;
+use vektor::x86::*;
 
 const PART_MASK: [u8; 128] = [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
                               0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
