+++ src/stride.rs
@@ -10,6 +10,7 @@
 use crate::arch::current::vecs::*;
 use crate::iters::{SIMDIterable, SIMDIterator, SIMDArray, SIMDObject, UnsafeIterator, SIMDSized};
 use crate::std::iter::{Iterator, ExactSizeIterator, FromIterator};
+use crate::vecs::*;
 
 // For AVX2 gathers
 use crate::std::mem::transmute;
